6.1

+++ PrivateFrameworks/HomeSharing.framework/HSCloudArtworkInfoRequest.h
@@ -5,11 +5,11 @@
 @interface HSCloudArtworkInfoRequest : HSRequest {
 }
 
-+ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2 sessionID:(unsigned int)arg3;
-+ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2;
++ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2 sessionID:(unsigned int)arg3;
++ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2;
 
 - (id)canonicalResponseForResponse:(id)arg1;
-- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2 sessionID:(unsigned int)arg3;
-- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2;
+- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2 sessionID:(unsigned int)arg3;
+- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2;
 
 @end
